работает без активного словаря, вью работает только с dto, добавил билдер!!

+++ src/main/java/org/example/web/controllers/GlossControllerImpl.java
@@ -51,9 +51,13 @@ public class GlossControllerImpl implements GlossController {
     public String removeGlossary(GlossaryDto glossary) throws BookShelfLoginException {
         try {
             if (!glossary.getName().isEmpty()) {
-                service.getGlossary(glossary.getName());
-                service.deleteGlossary(glossary);
-                return "redirect:/glossaries/shelf";
+                if (service.listAllWords().isEmpty()) {
+                    service.getGlossaryDto(glossary.getName());
+                    service.deleteGlossary(glossary);
+                    return "redirect:/glossaries/shelf";
+                } else {
+                    throw new BookShelfLoginException("Удалите слова из словаря!", "/glossaries/shelf");
+                }
             } else {
                 throw new BookShelfLoginException("Ошибка ввода при удалении словаря!", "/glossaries/shelf");
             }
@@ -68,7 +72,7 @@ public class GlossControllerImpl implements GlossController {
         try {
             if (!name.isEmpty()) {
                 logger.info("glossaries glossary");
-                model.addAttribute("gloss", service.getGlossary(name));
+                model.addAttribute("gloss", service.getGlossaryDto(name));
                 model.addAttribute("word", new WordDto());
                 model.addAttribute("wordsList", service.listAllWords());
                 return "glossary_words";
@@ -84,7 +88,7 @@ public class GlossControllerImpl implements GlossController {
     public String addWordValue(WordDto word, Model model) throws BookShelfLoginException {
         if (!word.getName().isEmpty() && !word.getValue().isEmpty()) {
             service.saveWord(word);
-            model.addAttribute("gloss", service.getGlossary(word.getGlossary()));
+            model.addAttribute("gloss", service.getGlossaryDto(word.getGlossary()));
             model.addAttribute("word", new WordDto());
             model.addAttribute("wordsList", service.listAllWords());
             return "glossary_words";
@@ -97,9 +101,9 @@ public class GlossControllerImpl implements GlossController {
     public String removeWordValue(WordDto word, Model model) throws BookShelfLoginException {
         try {
             if (!word.getName().isEmpty()) {
-                service.getWord(word);
+                service.getWordDto(word);
                 service.deleteWord(word);
-                model.addAttribute("gloss", service.getGlossary(word.getGlossary()));
+                model.addAttribute("gloss", service.getGlossaryDto(word.getGlossary()));
                 model.addAttribute("word", new WordDto());
                 model.addAttribute("wordsList", service.listAllWords());
                 return "glossary_words";
@@ -119,7 +123,7 @@ public class GlossControllerImpl implements GlossController {
                 model.addAttribute("glossary", new GlossaryDto());
                 model.addAttribute("glossName", word.getGlossary());
                 model.addAttribute("searchWord", word.getName());
-                model.addAttribute("word", service.getWord(word));
+                model.addAttribute("word", service.getWordDto(word));
                 return "search_result";
             } else {
                 throw new BookShelfLoginException("Введите слово для поиска!", "/glossaries/glossary", word.getGlossary());
